cambio de BigDecimal a double

+++ msvc-viaje/src/main/java/org/example/msvcviaje/clients/CuentaClient.java
@@ -10,5 +10,5 @@ import java.math.BigDecimal;
 public interface CuentaClient {
 
     @GetMapping("/cuentas/saldo/{id}")
-    BigDecimal obtenerSaldo(@PathVariable Long id);
+    double obtenerSaldo(@PathVariable Long id);
 }
